Update the delete method in the twoThreeTree
This changes one of the fundamental errors in the deletion process and follows the guidelines
as provided in https://www.cs.princeton.edu/~dpw/courses/cos326-12/ass/2-3-trees.pdf.
Instead of simply switching a given deleted element with its predecessor, the tree
will now make changes based on the cases mentioned in the link.
- The delete method now has no empty leaf creating height imbalances
- The tree will balance correctly when any element is deleted